Refactor to share rendering logic.
