+++ SolutionName.PreviewApp/SolutionName.PreviewApp.csproj
@@ -69,6 +69,9 @@
     <None Include="Entitlements.plist" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\SolutionName\SolutionNameRenderer.cs">
+      <Link>SolutionNameRenderer.cs</Link>
+    </Compile>
     <Compile Include="Main.cs" />
     <Compile Include="AppDelegate.cs" />
     <Compile Include="ViewController.cs" />
