+++ SolutionName/ShowSolutionName.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using AppKit;
-using CoreGraphics;
-using Foundation;
+﻿using AppKit;
 using MonoDevelop.Ide;
 using MonoDevelop.Projects;
 
@@ -64,58 +61,9 @@ namespace DT.VS4Mac.SolutionName
             if (solutionName != _lastSolutionName)
             {
                 _lastSolutionName = solutionName;
-                RenderSolutionNameIcon(solutionName);
+                var newAppIcon = SolutionNameRenderer.CreateIconWithSolutionName(solutionName, _defaultImage);
+                NSApplication.SharedApplication.ApplicationIconImage = newAppIcon;
             }
-        }
-
-        private static void RenderSolutionNameIcon(string name)
-        {
-            if (!string.IsNullOrWhiteSpace(name))
-            {
-                const float margin = 4;
-                NSString text = (NSString)name;
-                var paragraphStyle = (NSParagraphStyle)NSParagraphStyle.DefaultParagraphStyle.MutableCopy();
-                paragraphStyle.Alignment = NSTextAlignment.Center;
-                var attributes = new NSStringAttributes()
-                {
-                    Font = NSFont.SystemFontOfSize(19, NSFontWeight.Regular),
-                    ForegroundColor = NSColor.White,
-                    ParagraphStyle = paragraphStyle
-                };
-                var textRect = new CGSize(_defaultImage.Size.Width - margin * 2, _defaultImage.Size.Height - 2 * margin);
-                var rect = text.BoundingRectWithSize(textRect, NSStringDrawingOptions.UsesLineFragmentOrigin, attributes.Dictionary);
-                var centerAdjustment = _defaultImage.Size.Width - rect.Width - 2 * margin;
-                rect.Offset(margin + centerAdjustment / 2, margin);
-                var brandedImage = NSImage.ImageWithSize(_defaultImage.Size, false, (dstRect) =>
-                {
-                    _defaultImage.Draw(dstRect);
-                    DrawBackgroundInRect(rect);
-                    text.DrawInRect(rect, attributes);
-                    return true;
-                });
-                NSApplication.SharedApplication.ApplicationIconImage = brandedImage;
-            }
-            else
-            {
-                NSApplication.SharedApplication.ApplicationIconImage = _defaultImage;
-            }
-        }
-
-        private static void DrawBackgroundInRect(CGRect rect)
-        {
-            var backgroundColor = NSColor.FromRgba(50, 50, 50, 240).CGColor;
-            var borderColor = NSColor.FromRgba(100, 100, 100, 240).CGColor;
-            nfloat radius = 6;
-
-            var context = NSGraphicsContext.CurrentContext.CGContext;
-            context.SaveState();
-            context.SetFillColor(backgroundColor);
-            context.SetStrokeColor(borderColor);
-            var path = NSBezierPath.FromRoundedRect(rect, radius, radius);
-            path.LineWidth = 2;
-            path.Stroke();
-            path.Fill();
-            context.RestoreState();
         }
     }
 }
